move EventSourcingModule to rabix-engine and fix imports after merge conflicts
