+++ rabix-backend-tes/src/main/java/org/rabix/backend/tes/model/TESTask.java
@@ -1,9 +1,10 @@
 package org.rabix.backend.tes.model;
 
-import java.util.List;
-
 import com.fasterxml.jackson.annotation.JsonCreator;
 import com.fasterxml.jackson.annotation.JsonProperty;
+
+import java.util.List;
+import java.util.Map;
 
 public class TESTask {
