+++ rabix-cli/src/main/java/org/rabix/cli/BackendCommandLine.java
@@ -27,27 +27,9 @@ import org.rabix.common.service.download.DownloadService;
 import org.rabix.common.service.upload.UploadService;
 import org.rabix.common.service.upload.impl.NoOpUploadServiceImpl;
 import org.rabix.engine.EngineModule;
-import org.rabix.engine.service.BackendService;
-import org.rabix.engine.service.BootstrapService;
-import org.rabix.engine.service.BootstrapServiceException;
-import org.rabix.engine.service.ContextRecordService;
-import org.rabix.engine.service.IntermediaryFilesHandler;
-import org.rabix.engine.service.IntermediaryFilesService;
-import org.rabix.engine.service.JobService;
-import org.rabix.engine.service.JobServiceException;
-import org.rabix.engine.service.SchedulerService;
-import org.rabix.engine.service.SchedulerService.SchedulerJobBackendAssigner;
-import org.rabix.engine.service.SchedulerService.SchedulerMessageCreator;
-import org.rabix.engine.service.SchedulerService.SchedulerMessageSender;
-import org.rabix.engine.service.impl.BackendServiceImpl;
-import org.rabix.engine.service.impl.BootstrapServiceImpl;
-import org.rabix.engine.service.impl.IntermediaryFilesLocalHandler;
-import org.rabix.engine.service.impl.IntermediaryFilesServiceImpl;
-import org.rabix.engine.service.impl.JobReceiverImpl;
-import org.rabix.engine.service.impl.JobServiceImpl;
-import org.rabix.engine.service.impl.SchedulerServiceImpl;
+import org.rabix.engine.service.*;
+import org.rabix.engine.service.impl.*;
 import org.rabix.engine.status.EngineStatusCallback;
-import org.rabix.engine.status.impl.DefaultEngineStatusCallback;
 import org.rabix.engine.store.model.ContextRecord;
 import org.rabix.engine.stub.BackendStubFactory;
 import org.rabix.engine.stub.impl.BackendStubFactoryImpl;
@@ -55,12 +37,19 @@ import org.rabix.transport.mechanism.TransportPlugin.ReceiveCallback;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import com.fasterxml.jackson.core.JsonProcessingException;
-import com.google.inject.AbstractModule;
-import com.google.inject.Guice;
-import com.google.inject.Injector;
-import com.google.inject.Scopes;
-import com.google.inject.TypeLiteral;
+import java.io.File;
+import java.io.IOException;
+import java.lang.reflect.InvocationTargetException;
+import java.net.URI;
+import java.net.URISyntaxException;
+import java.net.URL;
+import java.nio.charset.Charset;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.Paths;
+import java.text.DateFormat;
+import java.text.SimpleDateFormat;
+import java.util.*;
 
 /**
  * Local command line executor
