--- rabix-engine-store-event-sourcing/pom.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <parent>
-        <artifactId>rabix</artifactId>
-        <groupId>org.rabix</groupId>
-        <version>1.0.2</version>
-    </parent>
-    <modelVersion>4.0.0</modelVersion>
-
-    <artifactId>rabix-engine-store-event-sourcing</artifactId>
-    <name>rabix-engine-store-event-sourcing</name>
-
-    <dependencies>
-        <dependency>
-            <groupId>org.rabix</groupId>
-            <artifactId>rabix-bindings</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.rabix</groupId>
-            <artifactId>rabix-engine-store</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-        <dependency>
-            <groupId>org.rabix</groupId>
-            <artifactId>rabix-engine-store-postgres</artifactId>
-            <version>${project.version}</version>
-        </dependency>
-    </dependencies>
-</project>
